Revert "Modularize CRT"
This reverts commit 4f1ae4e5b13669ed6ee488472be51f563d5a5801.
