+++ .github/workflows/run-tests.yml
@@ -29,6 +29,7 @@ jobs:
     - name: Run tests
       run: |
         python scripts/ci/run-tests
+
     - name: codecov
       run: |
         rm tests/coverage.xml
