+++ tests/unit/auth/test_signers.py
@@ -429,6 +429,12 @@ class TestS3SigV4Auth(BaseTestWithFixedDate):
         self.assertNotEqual(sha_header, 'UNSIGNED-PAYLOAD')
 
 
+class TestCrtS3SigV4Auth(TestS3SigV4Auth):
+    # Repeat TestS3SigV4Auth tests, but using CRT signer
+
+    AuthClass = botocore.auth.CrtS3SigV4Auth
+
+
 class TestSigV4(unittest.TestCase):
     def setUp(self):
         self.credentials = botocore.credentials.Credentials(
@@ -616,6 +622,11 @@ class TestSigV4Resign(BaseTestWithFixedDate):
         self.auth.add_auth(self.request)
         self.assertEqual(self.request.headers.get_all('Authorization'),
                          [original_auth])
+
+
+class TestCrtSigV4Resign(TestSigV4Resign):
+    # Run same tests against CRT auth
+    AuthClass = botocore.auth.CrtSigV4Auth
 
 
 class BasePresignTest(unittest.TestCase):
@@ -879,6 +890,17 @@ class TestSigV4Presign(BasePresignTest):
         query_string = self.get_parsed_query_string(request)
         signed_headers = query_string.get('X-Amz-SignedHeaders')
         self.assertNotIn('content-type', signed_headers)
+
+
+class TestCrtSigV4Presign(TestSigV4Presign):
+    # Run same tests against CRT auth
+    AuthClass = botocore.auth.CrtSigV4QueryAuth
+
+    def setUp(self):
+        # Use CRT logging to see interim steps (canonical request, etc)
+        # import awscrt.io
+        # awscrt.io.init_logging(awscrt.io.LogLevel.Trace, 'stderr')
+        super().setUp()
 
 
 class BaseS3PresignPostTest(unittest.TestCase):
